% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerape-package.R
\docType{package}
\name{powerape-package}
\alias{powerape}
\alias{powerape-package}
\title{powerape: Power Analysis for Average Partial Effects}
\description{
Simulation-based power analysis in which the effect size of interest is
the average partial effect (APE) or the average interaction effect (AIE)
from a nonlinear model such as probit or logit, including
smallest-effect-size-of-interest (SESOI) powering via confidence-interval
criteria for detection, minimum-effect, and equivalence claims
(Riesthuis, 2024) and robustness sweeps over contextual assumptions
(Hancock and Feng, 2025).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jespernwulff.github.io/powerape/}
  \item \url{https://github.com/jespernwulff/powerape}
  \item Report bugs at \url{https://github.com/jespernwulff/powerape/issues}
}

}
\author{
\strong{Maintainer}: Jesper Wulff \email{jwulff@econ.au.dk}

}
\keyword{internal}
